initial refactor of Ans_Lv2WFSS to Lv2WFSSNIS and Lv2WFSSNRC
This solves the exponential time usage.
